move instance to new file
